Moved views into sms.py
