+++ corehq/apps/domain/views.py
@@ -708,6 +708,7 @@ class DomainSubscriptionView(DomainAccountingSettings):
 
     @property
     def page_context(self):
+        from corehq.apps.domain.views.sms import SMSRatesView
         return {
             'plan': self.plan,
             'change_plan_url': reverse(SelectPlanView.urlname, args=[self.domain]),
@@ -3024,66 +3025,6 @@ class DeactivateTransferDomainView(View):
     @method_decorator(login_required)
     def dispatch(self, *args, **kwargs):
         return super(DeactivateTransferDomainView, self).dispatch(*args, **kwargs)
-
-
-from corehq.apps.smsbillables.forms import PublicSMSRateCalculatorForm
-from corehq.apps.smsbillables.async_handlers import PublicSMSRatesAsyncHandler
-
-
-class PublicSMSRatesView(BasePageView, AsyncHandlerMixin):
-    urlname = 'public_sms_rates_view'
-    page_title = ugettext_lazy("SMS Rate Calculator")
-    template_name = 'domain/admin/global_sms_rates.html'
-    async_handlers = [PublicSMSRatesAsyncHandler]
-
-    @use_select2_v4
-    def dispatch(self, request, *args, **kwargs):
-        return super(PublicSMSRatesView, self).dispatch(request, *args, **kwargs)
-
-    @property
-    def page_url(self):
-        return reverse(self.urlname)
-
-    @property
-    def page_context(self):
-        return {
-            'rate_calc_form': PublicSMSRateCalculatorForm()
-        }
-
-    def post(self, request, *args, **kwargs):
-        return self.async_response or self.get(request, *args, **kwargs)
-
-
-class SMSRatesView(BaseAdminProjectSettingsView, AsyncHandlerMixin):
-    urlname = 'domain_sms_rates_view'
-    page_title = ugettext_lazy("SMS Rate Calculator")
-    template_name = 'domain/admin/sms_rates.html'
-    async_handlers = [
-        SMSRatesAsyncHandler,
-        SMSRatesSelect2AsyncHandler,
-    ]
-
-    @use_select2_v4
-    def dispatch(self, request, *args, **kwargs):
-        return super(SMSRatesView, self).dispatch(request, *args, **kwargs)
-
-    @property
-    @memoized
-    def rate_calc_form(self):
-        if self.request.method == 'POST':
-            return SMSRateCalculatorForm(self.domain, self.request.POST)
-        return SMSRateCalculatorForm(self.domain)
-
-    @property
-    def page_context(self):
-        return {
-            'rate_calc_form': self.rate_calc_form,
-        }
-
-    def post(self, request, *args, **kwargs):
-        if self.async_response is not None:
-            return self.async_response
-        return self.get(request, *args, **kwargs)
 
 
 class BaseCardView(DomainAccountingSettings):
