+++ corehq/apps/domain/views/sms.py
@@ -76,8 +76,12 @@ from corehq.apps.accounting.utils import (
     log_accounting_error, domain_has_privilege, is_downgrade
 )
 from corehq.apps.hqwebapp.async_handler import AsyncHandlerMixin
-from corehq.apps.smsbillables.async_handlers import SMSRatesAsyncHandler, SMSRatesSelect2AsyncHandler
-from corehq.apps.smsbillables.forms import SMSRateCalculatorForm
+from corehq.apps.smsbillables.async_handlers import (
+    SMSRatesAsyncHandler,
+    SMSRatesSelect2AsyncHandler,
+    PublicSMSRatesAsyncHandler,
+)
+from corehq.apps.smsbillables.forms import SMSRateCalculatorForm, PublicSMSRateCalculatorForm
 from corehq.apps.toggle_ui.views import ToggleEditView
 from corehq.apps.users.models import Invitation, CouchUser, Permissions
 from corehq.apps.fixtures.models import FixtureDataType
@@ -133,3 +137,59 @@ from corehq.apps.hqwebapp.signals import clear_login_attempts
 from corehq.apps.ota.models import MobileRecoveryMeasure
 import six
 from six.moves import map
+
+
+class PublicSMSRatesView(BasePageView, AsyncHandlerMixin):
+    urlname = 'public_sms_rates_view'
+    page_title = ugettext_lazy("SMS Rate Calculator")
+    template_name = 'domain/admin/global_sms_rates.html'
+    async_handlers = [PublicSMSRatesAsyncHandler]
+
+    @use_select2_v4
+    def dispatch(self, request, *args, **kwargs):
+        return super(PublicSMSRatesView, self).dispatch(request, *args, **kwargs)
+
+    @property
+    def page_url(self):
+        return reverse(self.urlname)
+
+    @property
+    def page_context(self):
+        return {
+            'rate_calc_form': PublicSMSRateCalculatorForm()
+        }
+
+    def post(self, request, *args, **kwargs):
+        return self.async_response or self.get(request, *args, **kwargs)
+
+
+class SMSRatesView(BaseAdminProjectSettingsView, AsyncHandlerMixin):
+    urlname = 'domain_sms_rates_view'
+    page_title = ugettext_lazy("SMS Rate Calculator")
+    template_name = 'domain/admin/sms_rates.html'
+    async_handlers = [
+        SMSRatesAsyncHandler,
+        SMSRatesSelect2AsyncHandler,
+    ]
+
+    @use_select2_v4
+    def dispatch(self, request, *args, **kwargs):
+        return super(SMSRatesView, self).dispatch(request, *args, **kwargs)
+
+    @property
+    @memoized
+    def rate_calc_form(self):
+        if self.request.method == 'POST':
+            return SMSRateCalculatorForm(self.domain, self.request.POST)
+        return SMSRateCalculatorForm(self.domain)
+
+    @property
+    def page_context(self):
+        return {
+            'rate_calc_form': self.rate_calc_form,
+        }
+
+    def post(self, request, *args, **kwargs):
+        if self.async_response is not None:
+            return self.async_response
+        return self.get(request, *args, **kwargs)
