+++ corehq/apps/hqwebapp/urls.py
@@ -1,7 +1,7 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 from django.conf.urls import include, url
-from corehq.apps.domain.views import PublicSMSRatesView
+from corehq.apps.domain.views.sms import PublicSMSRatesView
 from corehq.apps.settings.views import (
     TwoFactorProfileView, TwoFactorSetupView, TwoFactorSetupCompleteView,
     TwoFactorBackupTokensView, TwoFactorDisableView, TwoFactorPhoneSetupView,
